move properties to StudyView
