+++ src/main/java/seedu/address/logic/StudyView.java
@@ -4,6 +4,7 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 
 import java.util.List;
 
+import javafx.beans.property.ReadOnlyProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import seedu.address.logic.commands.DoneCommand;
 import seedu.address.logic.commands.Command;
@@ -17,13 +18,21 @@ import seedu.address.model.deck.Deck;
 public class StudyView implements ViewState {
     private final Model model;
     public final List<Card> listOfCards;
-    private final SimpleObjectProperty<Card> selectedCard = new SimpleObjectProperty<>();
     private final Deck activeDeck;
+    private Card currentCard;
+    private studyState currentStudyState = studyState.QUESTION;
+    private final SimpleObjectProperty<String> textShown = new SimpleObjectProperty<>();
+
+    public enum studyState {
+        QUESTION, ANSWER;
+    }
 
     public StudyView(Model model, Deck deck) {
         this.model = model;
         this.activeDeck = deck;
         listOfCards = deck.getCards().internalList;
+        generateCard();
+        setCurrentStudyState(studyState.QUESTION);
     }
 
     @Override
@@ -33,7 +42,7 @@ public class StudyView implements ViewState {
             case DoneCommand.COMMAND_WORD:
                 return new DoneCommand();
             default:
-                if (model.getCurrentStudyState() == Model.studyState.QUESTION) {
+                if (getCurrentStudyState() == studyState.QUESTION) {
                     return new ShowAnswerCommand();
                 } else {
                     return new GenerateQuestionCommand();
@@ -47,6 +56,38 @@ public class StudyView implements ViewState {
     }
 
 
+
+    public void setCurrentCard(Card card) {
+        currentCard = card;
+    }
+
+    public Card getCurrentCard() {
+        return currentCard;
+    }
+
+    public void setCurrentStudyState(studyState state) {
+        currentStudyState = state;
+    }
+
+    public ReadOnlyProperty<String> textShownProperty() {
+        updateTextShown();
+        return textShown;
+    }
+
+    public studyState getCurrentStudyState() {
+        return currentStudyState;
+    }
+
+    public void updateTextShown(){
+        String text =  (currentStudyState == studyState.QUESTION)
+                ? currentCard.getQuestion()
+                : currentCard.getAnswer();
+        textShown.setValue(text);
+    }
+
+    public void generateCard() {
+        setCurrentCard(activeDeck.generateCard());
+    };
 
 
 }
