+++ src/main/java/seedu/address/logic/commands/GenerateQuestionCommand.java
@@ -1,6 +1,7 @@
 package seedu.address.logic.commands;
 
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.StudyView;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
@@ -9,13 +10,14 @@ import seedu.address.model.Model;
  */
 public class GenerateQuestionCommand extends Command {
 
-    public static final String MESSAGE_SUCCESS = "Answer";
+    public static final String MESSAGE_SUCCESS = "Question";
 
     @Override
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
-        model.generateCard();
-        model.setCurrentStudyState(Model.studyState.QUESTION);
-        model.updateTextShown();
+        StudyView studyView = ((StudyView) model.getViewState());
+        studyView.generateCard();
+        studyView.setCurrentStudyState(StudyView.studyState.QUESTION);
+        studyView.updateTextShown();
         return new CommandResult(String.format(MESSAGE_SUCCESS));
     }
 }
